backend/requirement.txt
fastapi
uvicorn[standard]
pydantic
requests
python-dotenv
numpy
earthengine-api